Separate user model from app.py
