+++ server/app.py
@@ -5,10 +5,10 @@ from api.home_handler import home_handler
 from config import DB_USERNAME, DB_PASSWORD, DB_NAME
 from flask_bcrypt import Bcrypt
 from email_validator import validate_email
+from models import bcrypt, User
 import jwt
 import json
 import re
-
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://{}:{}@localhost/{}'.format(DB_USERNAME, DB_PASSWORD, DB_NAME)
@@ -16,19 +16,6 @@ app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['SECRET_KEY'] = 'bubblegum'
 db= SQLAlchemy(app)
 bcrypt = Bcrypt(app)
-
-class User(db.Model):
-  __tablename__ = 'user'
-  id = db.Column(db.Integer, primary_key=True)
-  email = db.Column(db.String(120), unique=True, nullable=False)
-  password = db.Column(db.String(200), nullable=False)
-
-  def __init__(self, email, password):
-    self.email = email
-    self.password = bcrypt.generate_password_hash(password).decode('utf-8')
-
-  def __repr__(self):
-    return '<User %r>' % self.email
 
 @app.route('/signup', methods=['POST'])
 def signup():
@@ -46,7 +33,6 @@ def signup():
     return jsonify({'err': 'Account already exists'})
 
   token = jwt.encode(body, app.config['SECRET_KEY']).decode("utf-8")
-  print(token, type(token))
   return jsonify({'token': token})
   
 @app.route('/login', methods=['POST'])
@@ -58,7 +44,6 @@ def login():
       is_authenticated = bcrypt.check_password_hash(find_user.password, body['password'])
       if is_authenticated:
         token = jwt.encode(body, app.config['SECRET_KEY']).decode("utf-8")
-        print(token)
         return jsonify({'token': token})
       else:
         return jsonify({'err': 'Invalid credentials'})
